box: fix 64-bit schema version for message pack.
0b876b7600 introduced `uint64_t` schema version to deal with the
possible 32-bit counter overflow problem. But for some reason
message pack still serialized 64-bit schema version as 32-bit one.
Current commit fixes the issue.

NO_CHANGELOG=internal fix
NO_DOC=internal fix
NO_TEST=internal fix

+++ src/box/lua/net_box.c
@@ -2666,7 +2666,7 @@ netbox_transport_do_auth(struct netbox_transport *transport,
  * On success invokes the 'did_fetch_schema' callback and returns the actual
  * schema version. On failure raises a Lua error.
  */
-static uint32_t
+static uint64_t
 netbox_transport_fetch_schema(struct netbox_transport *transport,
 			      struct lua_State *L, uint64_t schema_version)
 {
@@ -2789,7 +2789,7 @@ restart:
  * Returns the current schema version on schema change. On failure raises
  * a Lua error.
  */
-static uint32_t
+static uint64_t
 netbox_transport_process_requests(struct netbox_transport *transport,
 				  struct lua_State *L, uint64_t schema_version)
 {
